Bug 669475 - Context menu item label setter only works for top-level items, not for items in submenus. r=myk
This is also practically a rewrite of context-menu.  It fixes a number of problems that occured when acting on menu items contained in submenus, such as setting their labels or destroying them.  It also adds addItem() and removeItem() to Menu.  Menu items now use light-traights and cortex.  Finally, it incorporates the fix to bug 671274.