Check if win exists before working on it
Sometimes I had an error, that win is null.
The error occured when I was logging in on a page where I was working
on. The login was done by an AJAX call but after the login page
reloaded. Then I tried to run my script I got the error message.

The error was come from the content/worker.js line 533